Moved selection changes and keyboard actions from DocumentComposer to DocumentInteractor.
